bugs fixed

+++ ft_calloc.c
@@ -6,14 +6,14 @@
 /*   By: josmarqu <josmarqu@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 09:20:35 by josmarqu          #+#    #+#             */
-/*   Updated: 2023/02/06 18:42:53 by josmarqu         ###   ########.fr       */
+/*   Updated: 2023/02/13 15:52:31 by josmarqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include "libft.h"
 
-void	*ft_calloc(int nmemb, size_t size)
+void	*ft_calloc(size_t nmemb, size_t size)
 {
 	unsigned char	*p;
